exercice exceptions

+++ src/main/java/com/adrar/exercice1/controller/LivreController.java
@@ -1,9 +1,11 @@
 package com.adrar.exercice1.controller;
 
+import com.adrar.exercice1.exception.LivreNotFoundException;
+import com.adrar.exercice1.exception.NoLivreFoundException;
 import com.adrar.exercice1.model.Livre;
-import com.adrar.exercice1.repository.LivreRepository;
 import com.adrar.exercice1.service.LivreService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
@@ -17,15 +19,21 @@ public class LivreController {
 
     @GetMapping("/books")
     public Iterable<Livre> getAllBooks() {
+        if(livreService.countLivre() == 0) {
+            throw new NoLivreFoundException();
+        }
         return livreService.findAllBooks();
     }
 
     @GetMapping("/book/{id}")
-    public Optional<Livre> getBookById(@PathVariable Long id) {
-        return livreService.findById(id);
+    public Livre getBookById(@PathVariable Long id) {
+        return livreService.findById(id).orElseThrow(
+                () -> new LivreNotFoundException(id)
+        );
     }
 
     @PostMapping("/book")
+    @ResponseStatus(HttpStatus.CREATED)
     public Livre addBook(@RequestBody Livre livre) {
         return livreService.addBook(livre);
     }
@@ -36,7 +44,11 @@ public class LivreController {
     }
 
     @DeleteMapping("/book/{id}")
-    public void deleteBook(@PathVariable Long id) {
+    public String deleteBook(@PathVariable Long id) {
+        if(!livreService.exists(id)){
+            return "Le livre n'existe pas";
+        }
         livreService.deleteBookById(id);
+        return "Livre supprimé avec succès";
     }
 }
